Split KubernetesCloud for easier management
