+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesCloudTest.java
@@ -1,20 +1,12 @@
 package org.csanchez.jenkins.plugins.kubernetes;
 
-import static org.junit.Assert.*;
+import java.util.Arrays;
 
 import org.csanchez.jenkins.plugins.kubernetes.volumes.EmptyDirVolume;
 import org.csanchez.jenkins.plugins.kubernetes.volumes.PodVolume;
 import org.junit.Test;
 
-import com.google.common.collect.ImmutableList;
-
-import java.util.Arrays;
-
 public class KubernetesCloudTest {
-
-    private KubernetesCloud cloud = new KubernetesCloud("test", null, "http://localhost:8080", "default", null, "", 0,
-            0, /* retentionTimeoutMinutes= */ 5);
-
 
     @Test
     public void testInheritance() {
@@ -31,33 +23,11 @@ public class KubernetesCloudTest {
         parent.setContainers(Arrays.asList(jnlp));
         parent.setVolumes(Arrays.asList(podVolume));
 
-
         ContainerTemplate maven2 = new ContainerTemplate("maven", "maven:2");
         PodTemplate withNewMavenVersion = new PodTemplate();
         withNewMavenVersion.setContainers(Arrays.asList(maven2));
 
         PodTemplate result = PodTemplateUtils.combine(parent, withNewMavenVersion);
-
-
-    }
-
-    @Test
-    public void testParseDockerCommand() {
-        assertNull(cloud.parseDockerCommand(""));
-        assertNull(cloud.parseDockerCommand(null));
-        assertEquals(ImmutableList.of("bash"), cloud.parseDockerCommand("bash"));
-        assertEquals(ImmutableList.of("bash", "-c", "x y"), cloud.parseDockerCommand("bash -c \"x y\""));
-        assertEquals(ImmutableList.of("a", "b", "c", "d"), cloud.parseDockerCommand("a b c d"));
-    }
-
-    @Test
-    public void testParseLivenessProbe() {
-        assertNull(cloud.parseLivenessProbe(""));
-        assertNull(cloud.parseLivenessProbe(null));
-        assertEquals(ImmutableList.of("docker","info"), cloud.parseLivenessProbe("docker info"));
-        assertEquals(ImmutableList.of("echo","I said: 'I am alive'"), cloud.parseLivenessProbe("echo \"I said: 'I am alive'\""));
-        assertEquals(ImmutableList.of("docker","--version"), cloud.parseLivenessProbe("docker --version"));
-        assertEquals(ImmutableList.of("curl","-k","--silent","--output=/dev/null","https://localhost:8080"), cloud.parseLivenessProbe("curl -k --silent --output=/dev/null \"https://localhost:8080\""));
     }
 
 }
